Lambs carcass measurements and tissues proportion

The dataset lambBreeds contains the information from 40 male lambs (22 of Churra Galega Bragançana (CGB) breed and 18 of Suffolk (SU) breed), carcass dimension measurements, tissues depth measurements, longissimus muscle area, breast bone tissue thickness, and carcass composition.

Usage

data(lambBreeds)

Format

A data frame with 40 rows (lambs) and 25 columns (variables):

Lamb

Lamb identification number

Breed

Breed of the lamb: 22 of Churra Galega Bragançana breed (CGB) and 18 of Suffolk breed (SFK)

HCW

Hot carcass weight (Kg) obtained imediatly after slaugther

F

Leg length (mm) - the smallest distance from the perineum to the interior face of the tarsal-metatarsal articular surface

K

Carcass length (mm) - measured from de base of the tail to the base of the neck

G

Width of the buttocks (mm) - the width measured using the measuring calliper at the level of the proximal edge of the patellae

U

Circumference of thorax (mm) - the circumference measured using a tape held horizontally around the thorax at the level of the caudal portion of the scapula

CB

Circumference of buttock (mm) - the circumference measured using a tape held horizontally around the buttocks at the level of the caudal insertion

C1

Subcutaneous fat thickness (mm) between the 1st-2nd lumbar vertebrae

C3

Subcutaneous fat thickness (mm) between the 3rd-4th lumbar vertebrae

C12

Subcutaneous fat thickness (mm) between the 12th-13th ribs

BT2

Total breast bone tissue thickness (mm) measured at middle of the 2nd sternebrae

BT3

Total breast bone tissue thickness (mm) measured at middle of the 3rd sternebrae

BT4

Total breast bone tissue thickness (mm) measured at middle of the 4th sternebrae

B1

Longissimus muscle depth (mm) measured between the 1st-2nd lumbar vertebrae

B3

Longissimus muscle depth (mm) measured between the 3rd-4th lumbar vertebrae

B12

Longissimus muscle depth (mm) measured between the 12-13th ribs

LEA1

Longissimus muscle area (\(cm^2\)) measured between the 1st-2nd lumbar vertebrae

LEA3

Longissimus muscle area (\(cm^2\)) measured between the 3rd-4th lumbar vertebrae

LEA12

Longissimus muscle area (\(cm^2\)) measured between the 12-13th ribs

MUS

Carcass muscle proportion (%)

SF

Carcass subcutaneous fat proportion (%)

IF

Carcass intermuscular fat proportion (%)

BO

Carcass bone proportion (%)

KKCF

Kidney-knob channel fat proportion (%)

References

CadavezMSci2012pasagropasr

Author

Vasco Cadavez vcadavez@ipb.pt

Examples

data(lambBreeds)
str(lambBreeds)
#> 'data.frame':	40 obs. of  25 variables:
#>  $ Lamb : num  9810 9819 9853 9902 9905 ...
#>  $ Breed: chr  "CGB" "CGB" "CGB" "CGB" ...
#>  $ HCW  : num  14.48 13.86 9.06 9.98 10.19 ...
#>  $ F    : num  279 297 269 296 262 301 297 280 297 280 ...
#>  $ K    : num  745 753 656 730 690 799 809 770 802 735 ...
#>  $ G    : num  222 207 191 192 192 207 196 210 210 198 ...
#>  $ U    : num  651 638 561 607 582 ...
#>  $ CB   : num  598 567 504 495 500 567 536 580 559 513 ...
#>  $ C1   : num  2.7 3.2 2.8 1.3 0.9 1.5 1 1.4 1 0.5 ...
#>  $ C3   : num  5.4 4.1 2.3 2.15 4 2.6 2 3.35 2.45 1.2 ...
#>  $ C12  : num  2.3 2.1 2.5 1.5 1.45 1.4 1.2 1.7 1 0.5 ...
#>  $ BT2  : num  24.3 21.7 18.5 17.3 16.8 ...
#>  $ BT3  : num  20.4 21.7 17.6 16.3 15.9 17 20.3 20 18 13.6 ...
#>  $ BT4  : num  17.1 18.9 15.4 12.9 11.9 ...
#>  $ B1   : num  27.7 24.5 23.7 22.1 26.8 ...
#>  $ B3   : num  26.5 22.2 20.1 22.7 17.5 24.3 25.5 23.3 26.4 21 ...
#>  $ B12  : num  26.4 23.5 23.6 20.3 26.4 ...
#>  $ LEA1 : num  11.4 10.1 9.6 7.9 9.05 ...
#>  $ LEA3 : num  14.2 13.6 9.4 9.05 8.75 ...
#>  $ LEA12: num  12.8 10.75 10.1 7.15 9.7 ...
#>  $ MUS  : num  59.1 59.3 58.4 62.2 58.9 ...
#>  $ SF   : num  7.21 8.07 8.24 4.59 6.59 6.37 5.48 7.22 5.68 4.54 ...
#>  $ IF   : num  10.91 8.84 9.6 5.91 8.34 ...
#>  $ BO   : num  20.7 21.6 21.9 26.1 24.4 ...
#>  $ KKCF : num  2.1 2.11 1.83 1.18 1.79 1.92 1.77 1.93 1.51 1.32 ...